Fix missing blockheight and incorrect RPC endpoint (#70)
There are two subquery examples for Boba.  One is for Boba BNB, while
the other is for Boba Ethereum.  Both have the same mistakes in the
implementation of the mappingHandlers, as both needlessly await on
already populated fields and additionally forget to populate the
(required) blockHeight field.

Additionally, the Boba Ethereum example re-uses the RPC endpoint for
Boba BNB.  Attempting to execute the example will result in a failure
due to a chain ID mismatch.

I'm not entirely sure whether this files are the source of, or product
of, the `subql init` step, so this might be the wrong spot to submit the
PR, if so, please redirect me and I'll be happy to adapt it.

Additionally

https://academy.subquery.network/quickstart/quickstart_chains/boba-bnb.html#add-a-mapping-function

and

https://academy.subquery.network/quickstart/quickstart_chains/boba-eth.html#add-a-mapping-function

likewise need to be regenerated/corrected.